after the newest version of interface
eṉ something (my something) is fixed now in the form eṉ+ something. 
other geminations are stilled marked with red. I changed marking this type of gemination: oḷ +iḻai > oḷ+ iḻai (just like I did with eṉ +something > eṉ+ something), but it's still red. 
However I feel that to mark the gemination at the beginning of the next word could be confusing, I would like transfer them to the end of the first word, like in the case of eṉ. I don't want though to change Eva's markup permanently, so for the time being, I left the + at the beginning of the second word.